requirements.txt
flask==2.3.2
requests==2.31.0
python-dotenv==1.0.0
fredapi==0.5.1
openai==1.3.0